requirements.txt
BeautifulSoup4
django-markdownify
crispy-bootstrap5